Implement SSR for ValidationGroup and ValidationSummary
