+++ testing/tests/DevExpress.serverSide/validationGroup.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/validationGroup.markup.tests.js");
