+++ testing/tests/DevExpress.serverSide/validationSummary.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets.editors/validationSummary.tests.js");
